requirements.txt
aerich==0.7.2
aiofiles==0.5.0
aiomysql==0.2.0
aiosqlite==0.16.1
aniso8601==7.0.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==4.0.3
asyncpg==0.29.0
atlastk==0.13.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
ciso8601==2.3.1
click==8.1.7
dictdiffer==0.9.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
importlib-metadata==6.8.0
iniconfig==2.0.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.3
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pip==23.3.1
pluggy==1.3.0
promise==2.3
psycopg2==2.9.9
pyasn1==0.5.1
pydantic==2.5.2
pydantic_core==2.14.5
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
PyJWT==2.8.0
PyMySQL==1.1.0
PyPika==0.48.9
pypika-tortoise==0.1.6
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2020.5
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
Rx==1.6.3
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.3
tortoise==0.1.1
tortoise-orm==0.20.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
watchgod==0.8.2
websockets==12.0
Werkzeug==3.0.1
wheel==0.38.4
zipp==3.17.0
bcrypt==4.1.1
